Create README.md
HOSPITAL MANAGEMENT SYSTEM
Here the website is used by patients, doctors and admin.
Patients are used to login and fix an appointment with the doctor and see the description what the doctor gives. 
Doctors can prescribe what to eat when the patient is unwell. If the doctor is not able to attend the patient, they can cancel the schedule.